gRPC: return nil payload when LocalCodeIntel fails (#47922)
Our client code expects a nil payload when LocalCodeIntel fails. This
seems to be used as a sentinel to fall back to search-based code intel
in the client code. For the gRPC implementation, we propagated any
errors we received, but this actually ends up breaking things because
errors cause the client to not fall back to search-based code intel and
instead throw the error.

@sourcegraph/code-intel, this feels like a hack. This probably isn't the
best way to solve this, but I don't want to get distracted right now.
Feel free to follow up with a less hacky fix.

## Test plan

Tested locally that I would get RPC errors showing up in the client when
the `grpc` flag was enabled, but applying this fix let the client fall
back to search-based code intel.