Return correct result counts for ownership search
Ownership search is implemented as a post-filter so we might drop results. Search usually expects every result to be returned, so we only request `N | count:N` results from each search backend. With filtering in place, though, we might not return the correct result count.
As a temporary fix, we're overwriting the result limit for search backends to `count:all`, then filter the results, and then rely on the limit job to correctly cut the stream once enough results were collected.