Add useSetting hook (#47979)
Inspired by Vova's #47864, this adds a new `useSettings()` hook to
replace the existing `SettingsCascadeProps` prop drilling.

I decided to _not_ replace all of the props drilling just yet to avoid a
massive merge conflict when Vova's PR ships. I also really want to focus
on cleaning up the app shell instead and I need this particularly for
one use case: Cleaning up `globbing` which, as it turns out, is just a
computed value from the setting instead.

With this new settings API in place, I want to start trimming the shared
props mess in a follow-up PR. This will then already remove the needs to
pass the settings cascade around naturally and we will end up with a few
smaller issues instead of a high change of breaking things.

I've also marked the existing settings APIs (including access to the
`SettingsCascade` object directly) as deprecated. `useSettings()` should
be used when possible. I don't think we need to validate the settings in
every caller manually.