experimental search input: Add toggle to enable/disable new search input
This PR adds a toggle button to the user menu to enable/disable the new
search input.

This changes the purpose of the experimental feature flag: If it's set
to `experimental` then the toggle button will be shown and the new
search input will be enabled by default.

The user can then enable/disable the new input via the toggle. This
setting is stored in temporary settings.

When the experimental feature flag is *not* set, neither the new search
input nor the toggle button will be available.

In order to roll out this change to customers we need to set the
experimental feature flag to `experimental` by default.

There is the caveat that if the user disabled the new search input it
will still be rendered briefly until the temporary settings have been
loaded. Not sure if there is a good way around this without blocking
rendering of the search input until the data is available (but I guess
we could do that if we wanted to).