experimental search input: Correctly handle clicks on "Go to" (#48036)
Currently the "Got to" (or secondary) action is only "selectable" via
keyboard shortcut. This commit updates the suggestions list
implementation to recognize clicks on the "Go To" label.

I've considered whether those should actually be links or buttons, but
I'm not sure it makes sense in this context because they are not
actually focusable anyway (real focus always has to stay inside the text
input), and keyboard events are handled by CodeMirror directly. I'll do
a proper a11y evaluation later.

To make it clearer which "action" is taken I'm adding a text-decoration
effect when hover over actions. I also tried to add the effect to the
primary action when hovering elsewhere over the option, but I couldn't
make it work properly.

## Test plan

Manual testing.