Fix wildcard tooltip making `<svg>` elements appear in tab order
As part of #48035, we observed an issue that only reproduces in Chrome.

When an `<svg>` element that should not be focusable is getting a focus
event listener attached, Chrome will _sometimes_ make it appear in the
tab order.

To reproduce this issue, you can take a look at this JSBin example:
https://jsbin.com/yucutapufi/edit?html,js,console,output

We found the respective bug report in the Chromium bug tracker:
https://bugs.chromium.org/p/chromium/issues/detail?id=445798

For now, we decided that we value browser consistency and are fixing
this by not attaching a focus event listener to `<svg>` elements that
do not have an explicit tabIndex property set to a value other than -1.

This PR _might_ make it so that some tooltip targets are no longer
accessible via tab navigation. However, these never were accessible in
other browsers and we would need to fix them regardless. This will just
make the problem more visible to us.