experimental search input: Support complex input updates
Our current search input allows for some "complex" input updates:
External places, like the search side bar or the exmaple chips can
update the query input and instruct it to e.g. focus itself or open
completions.

Until now the new query input did not support this. This PR extracts the
logic to perform these updates from the existing query input into a
reusable hook.

While doing that I also updated the new search input to reuse some of
functions I originally created for CodeMirror. And while doing _that_ I
noticed a bug in the original implementation that caused the input value
not to be updated. Turns out that you should listen to lint warnings...

But even with this fix in place I ran into problems with trying to
initialize the editor fully on first render and only trigger extension
reconfigurations when the extensions changed afterwards. In the end I
had to initialize the editor partially and reconfigure the extensions
when the editor becomes available and on extension changes.