gRPC: use global connections for gitserver (#47927)
Best practice for gRPC connections is to reuse them across many
requests. Currently, we are creating a new connection per RPC call. This
is bad, but it was the easiest way to see things running end-to-end.

This PR builds off https://github.com/sourcegraph/sourcegraph/pull/47924
to store a set of connections along with the current set of addresses.
The connections are updated to match the addresses every time the config
changes, and these updates are atomic, so there will never be a
situation where the connection list and the address list do not agree.

This is an alternative to using gRPC's client-side load balancing to
pick connections. To use gRPC load balancing, we'd need to communicate
the pick key (repo) via contexts, which is awkward. Additionally, it
would have to be somewhat leaky because we have many clients that depend
on the raw values of the addresses, and even some that do fan-out
requests that would need to hit all gitservers. Given all this, I think
it's best to stay closer to our current pattern.

## Test plan

Added tests for the gRPC server selection. All existing tests are
updated to exercise the new pattern.

---------

Co-authored-by: Geoffrey Gilmore <geoffrey@sourcegraph.com>