build: move build-config specific dependencies to build-config/package.json (#48072)
Other packages within `client/*` depend on `@sourcegraph/build-config`
as a dev dependency, but `@sourcegraph/build-config` itself depends on
these as runtime dependencies. Under bazel this dev vs non-dev is
stricter and `build-config` needs to declare the dependencies correctly.

Some of the deps are still in the root package.json as dev-deps when
they are still used elsewhere, others were removed from the root and are
now only in `build-config`.