job: remove result filter from keywordSearchJob
I believe this filter was added to increase the relevance of
keyword-search results. However, in my tests I have seen that it
is too aggresive and sometimes doesn't yield any results.

For example:

On .com this yields results:
```
zoekt query patternType:keyword r:sourcegraph$
```

but this doesn't
```
zoekt query asdf patternType:keyword r:sourcegraph$
```

This is not intuitive because we would expect the keyword search to be
robust against non-matching terms, IE the result set should not decrease
with more terms.

I propose to rely on our existing ranking (provided by Zoekt and
document ranking) and to remove the logic in the job layer.

This will make it much easier to reason about relevance for keyword-search.