service: ensure we always call ready once
The interactions with ready seemed like a footgun. For example it is
easy to forget to call ready which means we would never mark the process
as ready. In fact the executor service never called ready. Alternatively
you may accidently call ready twice leading to us marking the service as
ready too soon.

Additionally I cleaned up a ready call that wasn't needed anymore and
added documentation on the invariants we expect implementations to know.

Test Plan: go test. Additionally I audited each use of ready to ensure
it was called before the Start function returned.