app: open browser to url
This reduces the friction of needing to copy paste the sourcegraph URL
into your browser. When starting sourcegraph app we expect the users
intention is to use it in the browser.

We introduce the AppReady background job to frontend for this logic. I
believe this logic will be extended further for automatic account
creation/etc which is why it lives here. Open to suggestions for a
cleaner place.

Test Plan: sg start app and the browser is opened.