keyword: expand query with stems
Instead of replacing the original keywords with their stems, we keep
both.

Example:

Before:
I, like, to, query, Zoekt -> queri, zoekt ("I" and "like" are stop words)

After:
I, like, to, query, Zoekt -> query, queri, zoekt

This has 2 advantages:

1. More matches: The stem "queri" might not match, but the original
keyword "query" might.

2. Matching symbols: If a user query contains a symbol name, we want to
make sure that we match it extactly and give it a proper boost.

The change makes the query more expensive, however this is probably
fine, because we limit keyword queries to a single repo for now.

Test plan:
- updated unit tests