jscontext: pre-load CurrentUser by adding it to jscontext. (#47010)
On every web application page refresh, we send a GraphQL query to get
the current user. This query blocks the rendering of the react tree at
the very top of the hierarchy — in the `SourcegraphWebApp` component.
The query is relatively light, and we spend most of our time on a round
trip back to the server. To speed things up, we can inject this data
into the generated `index.html` page or, more precisely, into the
`window.context` object that we already use to propagate data from the
server to the client on page load.

We use the same GraphQL resolvers used for the `CurrentAuthState` query,
which means that the result is identical, and we can use it as a drop-in
replacement on the first load. This removes the initial round-trip to
the server required to start application rendering!