goroutine: inline StartableRoutine into BackgroundRoutine
I'm not sure why this is its own interface, maybe in the past it was
used? But from a diff search I didn't see any reasonable uses. There are
no references to this type so we might as well inline its definition
into the only place it is mentioned.

Test Plan: go test