add optional flag to skip pagination arg validation (#48101)
There are some resources we want to optionally pagination, for example,
RBAC permissions. When a user loads the Sourcegraph web app, we need to
fetch all permissions assigned to that user to figure out access
control.
This PR adds a field to `ConnectionResolverOptions` called
`SkipArgsValidation`, which skips the check to ensure `First` or `Last`
is set when accessing a paginated resource. The field is false by
default to enforce pagination for new and existing resources and I've
added a comment so anyone making use of it understands the performance
implications of setting it to `true`.



## Test plan

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->
* Manually tested
* Added unit tests