Search: do not write after starting stream (#48132)
After we make our first write to the stream, setting the HTTP status
code has no effect. Additionally, clients expects the response to be an
SSE-formatted stream, and writing the error message to the stream breaks
that expectation, giving us garbage errors.

This just removes the HTTP error status and depends on writing the error
to the stream as an event.