Ownership resolvers for blob view (#48056)
Part of #47980

This pull request implements just enough ownership resolver to support
the blob view.

I verified this manually in `sg start`. The added unit test mimics
exactly the behavior of my dev instance.

## Test plan

A single test case exercises a graphQL request made by the blob view in
case where we cannot resolve the owner.

There is an interesting case that I omitted, which is testing with the
Person resolver:

- this will likely require some faking behind the Person resolver (on
the database likely)
- that means more testing code
- so to keep this pull request size in check, let me do that in a
follow-up pull request.