repos: add ListRepos tests for Other
We had some tests specific to src-expose. This adds more generic tests
which help exercise a large amount of the logic. We don't assert on the
details of the repos type, this should be for another test. For now by
just testing on names we have an easier to read and reason about tests.

This is added since we intend on adding exclude support.

This commit additionally explicitly documents that repositoryPathPattern
is ignored for src-expose/src-serve.

Test Plan: go test

plz-review-url: https://plz.review/review/19564