replace parallel with internal limiter package
The parallel package is being used here as an unnecessarily complex
semaphore. It has all sorts of stuff like errors and waitgroups that we
don't need for these use cases, and what we do need is implemented
simply on top of a channel without adding a depenedency. Additionally,
the property that the zero value of the new limiter is usable cleans up
the callsites a bit.