web: review and better document the app shell context fields (#48172)
We have many fields coming from the tip of the React tree. It's hard for
me to understand where they come from because all of them are defined in
one huge interface. Understanding these fields' nature is essential
because we must consolidate all dynamic fields on one layer and update
them together to avoid redundant content layout shifts.

We also need to define guidelines for new application configuration
fields to make them predictable and easy to find. With this PR, I went
over all the fields in the application shell and split them into dynamic
and static categories. I also documented ideas on what we can do to
remove most of the dynamic fields from the application shell.