repoupdater: use localhost if deploy type is app (#48177)
We use env.Get at init() time for the repo updater URL. This doesn't
give enough time for app's setup to set the variable. So instead we
switch to deciding the value based on deploy type. This is the same
strategy used by gitserver, symbols, searcher and zoekt.

Test Plan: sg start app and CI