rbac: make pagination optional for. permissions  (#48118)
This PR makes pagination optional for the permissions resolver. We need
to be able to fetch a user's or role's permission in one query for the
following reasons:

* so that the background job for updating permissions can make just one
query to the database to fetch all permissions to compare. We previously
had a different method, `FetchAll`, but I removed it because it
duplicated the logic for `List` on the permission store.
* We need to be able to fetch a user's permission at startup to figure
out what aspects of the UI they have access to or should not have access
to

## Test plan

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->
* Manual testing.
![CleanShot 2023-02-23 at 15 26
14](https://user-images.githubusercontent.com/25608335/220935318-0c31ba38-0905-4073-80e6-272306c48216.png)