experimental search input: Fix history mode
Currently selecting a history entry from the search input when rendered
in the navbar leaves the input in a broken state.

I didn't notice this because when I added the history mode the input was
only rendered on the search home page. To make it work in other places I

- updated the extension to properly update the editor state (clear mode
  and set input)
- updated the global search query store to allow overwriting the query
  via parameters

I first attempted to remove the query parameter and instead relied on
the fact that updating the input value will update the query state, but
it seems due to effects this would try to submit the query before the
query state was updated.
Allowing the query to be set on submission via parameters seems the
safer solution.