auth/azureoauth: Use AZURE_DEV_OPS_API_URL as the ServiceID
When creating an external account we want to use AZURE_DEV_OPS_API_URL
as the ServiceID when storing an entry in the user_external_accounts
table because this is the URL used in the external service.

We're not able to use s.ServiceID in the assignment because it would
incorrectly assign the value of VISUAL_STUDIO_APP_URL to the service ID.
The result of this would then be:

During the permission sync, the ADO authz provider created uses
AZURE_DEV_OPS_API_URL as the service ID. But the entry in the
user_external_accounts_table would have the value of
VISUAL_STUDIO_APP_URL as the ServiceID. Leading to a mismatch and the
permissions syncing code falsely thinking that no external account
exists for this authz provider.

So why even use the VISUAL_STUDIO_APP_URL as the service ID in the auth
provider? This is because that's the URL that is used for all oauth2
requests. While the API for product features like repos, orgs, user
profiles is pointed to by the value of AZURE_DEV_OPS_API_URL leading to
a deviation in expectations.