schema: Add property enforcePermissions to ADO code host schema
Instead of continuing to use the `authorization` pattern similar to
other code hosts, it is simpler to add a boolean which has an explicit
name and describes what it does better.

The end result: we start to break away from the `authorization`
pattern (some other code hosts can do that as well, maybe this is the
start of it all) with an additional check for this new field in the
write path of external_services.Create, where we check to ensure the
correct value of `unrestricted` is written to the database.

For existing auth providers, this is forwards compatible. While at the
same time if they also wanted to get on the `enforcePermissions` way of
life in the future, the change to `recalculateFields` will let them be
backwards compatible.

Also a small tweak to the doc string to make things less confusing.