Search: disable ranking in hybrid search (#48164)
This PR proposes reverting #44647 for now. At the moment, we always enable
search ranking for hybrid search -- it is not controlled by the
`search-ranking` feature flag. Recently, we had an incident with a new ranking
change, and we couldn't fully disable the feature because hybrid search
hardcodes the flag to 'true'. While we're still actively developing search
ranking (and while it's still disabled by default), I think we should disable
it in hybrid too.