Remove `searchStats` feature flag and update docs (#47959)
This flag had been removed in #46045 but (accidentally?) added back in
#45705. The feature (search stats) itself been removed in #45996.
Documentation about it has already been removed in #30564. The link
removed in this PR has been dead since then.

## Test plan

`grep`ped the code for references to the flag and used sourcegraph to
find the commits that made changes to the related code and
documentation.