Backend: replace `lib/group` with `sourcegraph/conc` (#48162)
This replaces `lib/group` with `sourcegraph/conc`, which is its
spiritual successor.

## Test plan

Existing tests. `conc` itself is extremely well tested. The APIs of the
two packages are nearly identical since `conc` is based off `lib/group`.
The only significant change is that calling `Wait()` on a pool multiple
times is an error, so I fixed the couple places we do that.