Fix integration test
This was a weird issue, but it did uncover an actual problem.

Reason why test failed: It looks like the combination of focusing the
input plus updating the value caused Chrome (not Firefox) to have the
cursor be stuck at position 1, even though CodeMirror internally thought
it was at position 4.
When the test focused the input (again), it would enter the text at the
wrong position. Even manually clicking inside the input _after_ the
value did not move the cursor to the end of the value...

Now, either removing the focus behavior or not updating the value fixes
the problem.

This is possibly a bug in Chrome because Firefox doesn't show this
problem.

Uncovered actual problem: The input was focused becaused the "focus when
search context changes" logic was triggered. I believe this happens now
because the editor is intatiated before this effect runs. However it
should only run when the search context truly changed and not during the
initial render. I added `useOnValueChanged` to handle this case.

Additionally I changed the CodeMirror state update logic to only include
`changes` in the transaction when the values are different.
# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch fkling/refactor-codemirror-hook
# Your branch is up to date with 'origin/fkling/refactor-codemirror-hook'.
#
# Changes to be committed:
#	modified:   client/branded/src/search-ui/input/CodeMirrorQueryInput.tsx
#