web: Refactor useCodeMirror hook (#48193)
This was quite a journey: I set out to fix some issues with the new
query input and wanted to switch to using the `useCodeMirror` hook for
it at the same time... only to discover that the input was now broken in
a different way! Somehow the way that `useState` and `useEffect` had
been used didn't work as intended (specifically it was referencing the
wrong extensions object). I tried various solutions to no avail...

In the end I found [this example in the new React
docs](https://beta.reactjs.org/reference/react/useEffect#controlling-a-non-react-widget)
and decided to model the hook in a similar way.

The `useCodeMirrorHook` now received two refs: One for the editor and
one for the container DOM element. The editor ref can be used by the
caller to reference the editor in callbacks and effects.

When a component that uses `useCodeMirror` is rendered the first time
the following execution flows happens:
- Component registers "editor update effects" (because they are listed
before `useCodeMirror`). `useCodeMirror` is called with refs, initial
value and initial extensions. The hook registers its own update effects
and an editor creation effect.
  - React creates DOM nodes and sets container ref
  - React executes effects
- "Editor update effects" are run but don't have any effect because the
editor ref is still unset
- Editor creation effect is run, with access to initial value and
extensions. Creates editor and sets editor ref

This all worked great until I updated the `NotebookMarkdownBlock`
component, which renders an editor conditionally. But because the
container element is now passed as a ref, adding and removing the
container won't trigger the editor to be created/destroyed anymore.

For this case I introduced a simple `CodeMirrorEditor` component that
exposes a `focus` method. Being its own component, the editor gets
properly created/destroyed on conditional renders.

Despite the conditional rendering issue, overall I think it's the more
correct approach since it avoids unnecessary re-renders because the
container node is not stored in state anymore. The editor should now be
fully initialized in the first render pass.

## Test plan

- Typing a regular expression into the query input (e.g. `foo.*bar`) and
toggling the regexp button correctly updates the inputs syntax
highlighting (indicating that extensions get properly updated).
- Selecting a different search context focuses the query input

- In the file view, clicking the blame button toggles the blame view

- In Notebooks, clicking the render/edit button of a Markdown block
correctly switches between render and input mode.