Setup Experience: Fetch github repositories for remote github source (#48096)
The current setup of a code host connection of type Github expects the
user to provide criteria for matching remote repositories to be included
for mirroring/sync'ing with sourcegraph. For Github some of the ways to
define criteria are via `repositoryQuery` and with `repos` list.

The setup experience effort 
will fetch a list of github repositories so that user can view this list
of repositories and select which should be included in the code host
connection configuration. This fetch can take place before or after the
code host connection is saved to sourcegraph's External Services data
store. Criteria determining repository results is the `query` string,
and an empty query string will return all affiliated repos.
Additionally, one can pass `excludeRepos` list to ensure they are not
included in the results set.

## Test plan

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

add tests for github source, external services, and repoupdater.