Search: fix some `repo:contains.path()` bugs (#48215)
This commit fixes three issues:

1) a panic that only occurs when `index:no` is specified.
This is an undocumented debugging parameter, so it is not a panic we've
been seeing in production.

2) support for negating `repo:contains.path()` and
`repohasfile:` for unindexed repositories. When we added negation
support to that predicate, we missed the codepath that is run for
repositories that aren't indexed by Zoekt. Additionally, it extracts
some logic to make the flow a bit more clear.

3) the original issue, which is that we return "revision not
found" alert for repositories that have no commits. If a repository has
no commits, it definitely does not contain a file, so it's safe to
ignore the error in that case.