use types.ExternalServiceRepository instead of types.Repo (#48308)
Followup from https://github.com/sourcegraph/sourcegraph/pull/48096

This changes the ExternalServiceRepositories response to return
types.ExternalServiceRepository instead of types.Repo. We do not
actually use all the fields of types.Repo, and always convert it to a
types.ExternalServiceRepository. Additionally, a types.Repo cannot be
safely serialized to JSON because it loses the type information in the
Metadata field, so we don't want to have invalid types.Repo values
floating around.

This is blocking adding a gRPC implementation of these new RPCs because
I don't want to implement an incorrect protobuf encoding of
`types.Repo`.