own: Cache compiled rule expressions
This one is a bit controversial!
First, I've moved most custom code from the generated package. This is because I need to add additional fields to the struct definitions, and not just patch methods on top of it. This feels kinda clean to me as it separates app logic and serialization concerns into two different buckets, but let me know if you hate this and we can try to find another way.
Second, this is adding a sort of "hidden" caching layer here, that increases memory of a codeowners.File slightly over time as more rules have been touched.

Overall, this is giving quite a significant performance boost, because we don't need to recompile the patterns ever time we look at them, which for search can be thousands of times. I've also made sure that rules are lazy compiled the first time they're being used, so we only compile as many as we did before in this method vs pre-compiling them all.

Performance boost for calls to `FindOwners` calls after the cache is warm:
```
Before:
BenchmarkOwnersMatch-10           143442              8295 ns/op
After:
BenchmarkOwnersMatch-10           899032              1279 ns/op
```