Ranking: simplify final flush in aggregator
Before, when we received a first result from every endpoint, we incorrectly
flushed with the reason 'timer_expired'. This switches to 'final_flush', to
capture that the aggregator is done as it's seen results from all endpoints.

It also removes logic to perform a last 'flush' after all endpoints are done
streaming. This is actually a dead code path, since all endpoints will have
responded by then (through either calling Send or SendDone).