esbuild.context requires calling rebuild to build initially
I had assumed that `esbuild.context(...)` performs the initial build, but that was incorrect. You must call `rebuild()` on the awaited promise to perform the initial build.

Because esbuild is still only used as an alternative build system in local dev (and by occasional VS Code extension builds), this was not noticed before.