Highlighting: undo broken languages
Highlighthing is timing out for Ruby, C++, and Python, on
sourcegraph.com because we're trying to highlight these languages with
tree-sitter when syntax-highlighter has no tree-sitter support for those
languages. The grammars for these languages depend on C++ code that we weren't
able to build in CI so we removed them last minute before merging
the PR that added tree-sitter highligting for Java, TypeScript,
JavaScript, and Rust. https://github.com/sourcegraph/sourcegraph/pull/47571