Highlighting: undo broken languages (#48328)
Highlighthing is timing out for Ruby, C++, and Python, on
sourcegraph.com because we're trying to highlight these languages with
tree-sitter when syntax-highlighter has no tree-sitter support for those
languages. The grammars for these languages depend on C++ code that we
weren't able to build in CI so we removed them last minute before
merging the PR that added tree-sitter highligting for Java, TypeScript,
JavaScript, and Rust.
https://github.com/sourcegraph/sourcegraph/pull/47571

## Test plan

Manually tested that the blob view no longer times out after these
changes. We need to add regression testing for all the main languages to
prevent this kind of error from happening again.

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->