[feat] use unified permissions in authzQuery (#48263)
## Description

This PR modifies the authzQuery to read from the new unified permissions
table if the feature flag is enabled. We still read permissions from the
old table if no permissions are found in the new `user_repo_permissions`
table - in case the permissions were not written there yet.

Moved the functionality behind `experimentalFeatures.unifiedPermissions`
flag instead of dabatase feature flag that was used before. Making an
SQL request to get the value of the feature flag on every authzQuery
call would be wasteful and this way the value is cached in memory, so
much better performance wise. Also, easier to unit test.

## Test plan

Tested manually + updated unit tests