sveltekit: React integration (#48269)
This PR adds various helpers to integrate existing React routes into the
SvelteKit prototype. The idea is that we create pages which handle a
complete route "namespace" (e.g. `/notebooks/*`) delegate navigation
within this namespace to the React component / React router.

In order to properly handle "cross framework links" (i.e. a link
rendered with SvelteKit pointing to a React page and vice versa), and to
make navigation state synchronization more robust in general I've
introduced a custom React Router router which integrates with
SvelteKit's router API.

This PR adds support for the `/notebooks` and `/insights` routes but I
could only really test Notebooks because insights isn't available on
sourcegraph.com (which the prototype proxies to).

## Test plan

Run the prototype and navigate to, from ans withing `/notebooks`.