own: Improve performance of glob pattern matching (#48314)
Uses a small library that does ~what I was trying to build myself. It
reduces the complexity drastically by only supporting .* holes and being
very optimized for that. One side-note here is that it technically also
supports `?` characters. We can later fork this and/or write a similar
solution to adjust this to our needs.

Before and after for a medium complexity pattern `*RE*ADME.md`:

```
BenchmarkOwnersMatch-10          1204886               986.0 ns/op
BenchmarkOwnersMatch-10          4568581               258.1 ns/op
```