gRPC: fix SymbolInfo like LocalCodeIntel (#48377)
We previously made a fix for code
intel hovers in which we return nil and no error for cases when a
language isn't supported. However, we run into this same edge case with
the SymbolInfo endpoint. I've updated the logic to still return an error
over the RPC boundary, but then to interpret an "unimplemented" error as
no error and just return nil, similar to how the HTTP implementation
handles this.

This is currently breaking search-based codeintel on Rust because we
return an error rather than no results, so the UI bombs out rather than
falls back to search-based code intel.