Ranking: small fix to search aggregator (#48161)
This is a follow-up to the previous PR. It adjusts the logic so that we still
aggregate empty events. This way, we're careful to aggregate any stats from
these events.

The PR also improves a recently added test for the max memory limit.

## Test plan

I didn't add an explicit unit test for this. The core functionality covered by
`aggregate_test.go`.