httpcli: remove high cardinality prometheus label internal transport (#48417)
On sourcegraph.com the label with the highest memory usage is category
(562346 bytes vs the 2nd which is 179764). We tracked it down to the
metric src_internal_request_duration_seconds_bucket. For example we
register 3 metrics for every git repo we clone internally thanks to
using the path.

There is no reasonable alternative for us to use, nor is anyone relying
on this metric and category. So we are removing it for now by just
recording unknown.

Test Plan: go test

Co-authored-by: Stefan Hengl <stefan@sourcegraph.com>