servegit: use fastwalk from x/tools (#48435)
I've used this code before for my own personal projects, its super fast
for me. This won't solve all our problems yet, but atleast it is a lot
faster.

This commit is minimal change to the code to avoid too much whitespace
change. The next commit restructures the code a bit to make the code a
bit more robust.

Test Plan: go test