servegit: skip files unlikely to contain code
I tweaked this code until it ran in under 1s for my home directory on my
macbook. So all heuristics in it are based on that. They could easily
not apply for another user, so another commit will try and bound the
work done.

Test Plan: Ran locally via app-discover-repos

  $ go install ./dev/internal/cmd/app-discover-repos
  $ /usr/bin/time app-discover-repos -root ~ | wc -l
    0.43 real         0.78 user         0.94 sys
    43