servegit: introduce MaxDepth and Timeout to discovery
With these two parameters we should avoid overly long discovery
operations. Right now if we hit these we just log to the console a
warning which isn't the best API. But as we shape this API up we should
be able to better report these errors.

The default of 10 max depth and 5s is plucked out of thin air. Will
likely need tuning.

Test Plan: ran app-discover-repos -root ~ with the following different
values

- n/a :: everything worked as usual
- SRC_DISCOVER_TIMEOUT=10ms :: got timeout warning
- SRC_DISCOVER_MAX_DEPTH=3 :: only got repo names with at most 3 components