monitoring: add codeintel, batches multi-instance executor queue (#48305)
Adds multi-instance dashboards for each queue to Centralized
Observability, to allow us to monitor executor backlogs. The resulting
panels will look like this:
https://monitoring.sgdev.org/goto/hPiNMpxVz?orgId=1, but with one panel
for each of the queue metrics.

This is part of https://github.com/sourcegraph/customer/issues/1973 - we
don't have great visibility into the state of executor backlogs today.
The above example indicates it's not pretty - we seem to have backlogs
in the tens of thousands

This change includes:

1. Creating a new `shared.Executors`, moving
`shared.CodeIntelligence.NewExecutorQueueGroup` to it
2. Having `shared.Executors.NewExecutorQueueGroup` implicitly enable
multi-instance on its panels when used with a static queue filter
3. Add a `shared.Batches.NewExecutorQueueGroup` to the `worker` panel,
which seems to have most of the batches-specific panels (there isn't an
equivalent of the Code Intelligence dashboards for batch changes)

## Test plan

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

Generated output is the same. Manual testing of the multi-instanc
queries in https://monitoring.sgdev.org/goto/hPiNMpxVz?orgId=1