search input: Refactor "focused filter" logic
The current search input has an extension to decorate the "focused"
filter (the filter the cursor is at) to show a different background. The
same extension also showed a filter placeholder text when it was
available.

This PR refactors this logic into two separate extensions so the
placeholder one can be reused by the new search input (and maybe also
the (now called) active filter extension, but that's still up for
debate).

This PR also changes how these extensions work: Until now the extension
itself decided to render the decorations based on whether the input was
focused or not. Now this job is delegated to CSS.
What's great about this is that it also fixes an issue where the
CodeMirror input was somehow "stealing" focus when a filter was active
and the user tried to move the focus away from the input. win-win