Propose pattern to inject enterprise jobs into search pipeline (#48409)
With Own, we have a need to inject enterprise code into the search
pipeline. After a few different attempts, I have settled on this
approach. It does not feel _too_ invasive (well of course it does but
not much more than anything else in our annoying enterprise/oss split,
you know) and is fairly easy to adopt in other services outside of
frontend, too which is showcased in the worker job for codemonitors, for
example. I've made this non-own specific so in the future we can hook in
more jobs here. I've also implemented an unimplemented version (lol that
sounds weird) of it so that users properly get informed when a search
job type is not supported on the instance. Every job creator can decide
to return a NoopJob instead, but for Own we find it more natural than
instead trying to find files called `has.owner(erik)`.

Let me know what you think about this pattern! If this flies well, I'll
be able to move all the backend own code into enterprise after this
lands :)