Fix snippet formatting in human messages from recipes (#69) (#74)
And add an integration test checking that it shows up in the transcript.
Testing vscode extensions involves a bit of indirection: there's the
harness to run vscode, the test suite which runs in vscode, the
extension under test, and the content in the webview itself. All of
these are in isolated JavaScript contexts and often separate processes.

This first test plumbs through most of these layers but obvious
improvements are noted with TODOs.