search input: Refactor "focused filter" logic (#48482)
The current search input has an extension to decorate the "focused"
filter (the filter the cursor is at) to show a different background. The
same extension also showed a filter placeholder text when it was
available.

This PR refactors this logic into two separate extensions so the
placeholder one can be reused by the new search input (and maybe also
the (now called) active filter extension, but that's still up for
debate). The extensions use class names that can be used to customize
styles for different instances.

This PR also changes how these extensions work: Until now the extension
itself decided to render the decorations based on whether the input was
focused or not. Now this job is delegated to CSS.
What's great about this is that it also fixes an issue where the
CodeMirror input was somehow "stealing" focus when a filter was active
and the user tried to move the focus away from the input. win-win



## Test plan

- Enable old query input
- Type in filter with placeholder, e.g. `file:`: Filter has a different
background and placeholder shows.
  - Type 'space': Background color and highlight disappear
- Press left arrow to go back to filter: Highlight and placeholder
appears
  - Entering a value for the filter hides the placeholder
- Moving focus outside the input makes the background color and
placeholder disappear

- Enable new input
  - Typing `file:` shows placeholder
  - Entering a filter value hides the placeholder

## App preview:

-
[Web](https://sg-web-fkling-reuse-filter-placeholder.onrender.com/search)

Check out the [client app preview
documentation](https://docs.sourcegraph.com/dev/how-to/client_pr_previews)
to learn more.