bazel: add command sets that uses bazel under the hood (#48460)
This PRs introduces `bazelCommand`(s) in `sg.config.yaml` to handle
starting those services, plus command sets for both oss and enterprise.

`go run ./dev/sg start enterprise-bazel` will start most of the services
with bazel and will keep using normal commands for the others (caddy,
etc ...).

The code powering those is built on top of `startedCmd` which I don't
want to touch until we have a bit more understanding how all of this
behaves. Once we have that, we can start refactoring this so it's more
straightforward.

Running the command set with bazel is bit verbose, you'll see two new
"services" being ran:

- `[ bazel]` which is ran at the very beginning and the equivalent of
the old `install` part of the commands, except it does all of them at
once. Once it's complete, it will stop exit nicely.
- `[ iBazel]` which handles rebuilding the services in the background.
The only thing it will do is to watch the FS based on the bazel rules
and the depencency tree to understand what exactly need to be rebuilt.
Once it detects some changes, it will overwrite the existing binary that
was previously built by the first step (`[ bazel]`).

The bazel commands will detect that the binary has been overwritten and
will restart gracefully once it's done, meaning that the services will
still be running during the new build.

EDIT: added the codeintel command set for @efritz who's up to be our
guinea pig on this

## Test plan

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

Locally tested.

---------

Co-authored-by: davejrt <davetry@gmail.com>
Co-authored-by: William Bezuidenhout <william.bezuidenhout@sourcegraph.com>